Step 2/requirement.txt
pypdf
langchain
grobid_client_python

